requirements.txt
torch>=2.1.2
torchvision>=0.16.2
timm>=0.9.12
transformers>=4.37.0
accelerate>=0.25.0
pandas>=2.1.4
numpy>=1.24.4
scikit-learn>=1.3.2
albumentations>=1.3.1
Pillow>=10.1.0
opencv-python-headless>=4.8.1.78
easyocr>=1.7.1
pytesseract>=0.3.10
matplotlib>=3.8.2
seaborn>=0.13.0
tqdm>=4.66.1
streamlit>=1.29.0
lime>=0.2.0.1
shap>=0.44.0
captum>=0.6.0
fpdf2>=2.7.9
pytorch-grad-cam>=1.4.0
